+++ thetawave_lib/src/spawnable/entities/mod.rs
@@ -1,7 +1,4 @@
 mod blast;
 mod spawn;
 
-pub use self::{
-    blast::spawn_blasts,
-    spawn::{spawn_consumable, spawn_effect, spawn_item, spawn_mob, spawn_spawnable},
-};
+pub use self::{blast::spawn_blasts, spawn::spawn_spawnable};
